WaveAnimation - Changed `skew` to a signed byte to support reverse directions.

+++ src/animation/waveanimation.cpp
@@ -5,11 +5,11 @@ namespace PixelMaestro {
 		type_ = AnimationType::Wave;
 	}
 
-	uint8_t WaveAnimation::get_skew() {
+	int8_t WaveAnimation::get_skew() {
 		return skew_;
 	}
 
-	void WaveAnimation::set_skew(uint8_t skew) {
+	void WaveAnimation::set_skew(int8_t skew) {
 		this->skew_ = skew;
 	}
 
@@ -17,10 +17,10 @@ namespace PixelMaestro {
 		for (uint16_t y = 0; y < section_->get_dimensions()->y; y++) {
 			for (uint16_t x = 0; x < section_->get_dimensions()->x; x++) {
 				if (orientation_ == Vertical) {
-					section_->set_one(x, y, get_color_at_index(y + (cycle_index_ + (x * skew_))));
+					section_->set_one(x, y, get_color_at_index(y + (cycle_index_ + (int8_t)(x * skew_))));
 				}
 				else {	// Horizontal
-					section_->set_one(x, y, get_color_at_index(x + (cycle_index_ + (y * skew_))));
+					section_->set_one(x, y, get_color_at_index(x + (cycle_index_ + (int8_t)(y * skew_))));
 				}
 			}
 		}
